chore: move analytics related code outside of package and into mcp integration
