+++ typescript/src/integrations/mcp/utils/handleToolError.ts
@@ -1,5 +1,5 @@
+import { getPostHogClient } from "@/integrations/mcp/utils/client";
 import type { CallToolResult } from "@modelcontextprotocol/sdk/types.js";
-import { getPostHogClient } from "@/lib/client";
 
 export class MCPToolError extends Error {
 	public readonly tool: string;
